Optimize database path initialization logic
Deleted the logic of old database file copying.
Modified the initialization logic of `databasePath`, and now directly obtains the path from `AppSettings.DatabasePath`.
Added a check to see if `databasePath` is an absolute path. If it is not an absolute path, it will be converted to a path based on `AppContext.BaseDirectory`.