Updated message sending logic, added sender name parameter
In `GlobalMessage.cs`, updated the return value of `getDescription` and `getHelp` methods, added the description of `senderName` parameter. Updated the logic of `Execute` method, added the `senderName` parameter, and used the `Utils.CreateGameMessage` method to create messages.

In `SayToPlayer.cs`, updated the return value of `getHelp` method, added the description of `senderName` parameter. Updated the logic of `Execute` method, added the `senderName` parameter, and used the `Utils.CreateGameMessage` method to create messages.

In `ModEventHub.cs`, updated the parameter name of `OnChatMessage` method, changed from `mainName` to `senderName`, and added the processing logic for `senderName` being empty.