Fix chat message handling and improve code comments
In `GlobalMessage.cs`, the parameter `_senderEntityId` of the `ChatMessageServer` method is explicitly specified as `-1`.

In `SayToPlayer.cs`, add a check for the number of parameters, and if there are less than 2 parameters, log the help information and return. At the same time, adjust the initialization position of `senderEntityId`.

In `SayToPlayer.cs`, the parameter `_senderEntityId` of the `ChatMessageServer` method is explicitly specified as `senderEntityId`.

In `Common.cs`, change the comment from Chinese to English.

In `ModEventHub.cs`, add a reference to `using UnityEngine.Windows;`.
In `ModEventHub.cs`, the parameter `entityId` of the `OnChatMessage` method is renamed to `senderEntityId`, and `senderName` is renamed to `mainName`.
In `ModEventHub.cs`, the processing logic for `senderName` is added. If `clientInfo` is `ModApi.CmdExecuteDelegate`, `senderName` is extracted from `message`; if `senderEntityId` is `-1`, the localized server name is used; otherwise, `mainName` or `Common.NonPlayer` is used.
In `ModEventHub.cs`, the `EntityId` property of the `ChatMessage` object is changed from `entityId` to `senderEntityId`.