Fix message sending logic and optimize null value checking
In `FunctionBase.cs`, the `SenderName` property of `SendGlobalMessage` and `SendPrivateMessage` methods is changed from `ConfigManager.GlobalSettings.ServerName` to `ConfigManager.GlobalSettings.GlobalServerName` and `ConfigManager.GlobalSettings.WhisperServerName`.
In `GlobalMessage.cs`, the `SenderName` property is added.
In `Utils.cs`, the command format string of `SendGlobalMessage` and `SendPrivateMessage` methods adds the `SenderName` parameter.
In `Utils.cs`, the parameter type of the `FormatCommandArgs` method is changed from `string` to `string?`, and the null value checking logic is modified.
In `ServerController.cs`, a reference to `ConfigManager` is added.
In `ServerController.cs`, the `SendGlobalMessage` and `SendPrivateMessage` methods add a check to see if `SenderName` is empty before calling the `Utils` method, and set a default value if it is empty.