Added automatic zombie cleanup and auto restart features
Added `EnableAutoZombieCleanup` and `AutoZombieCleanupThreshold` settings in `functionsettings.json`.

Added several new properties and annotations in `GlobalSettings.cs`, including `IsEnabled` and `ExecuteCommands` properties of the `Trigger` class, and `IsEnabled`, `RestartHour`, `RestartMinute`, and `Messages` properties of the `AutoRestart` class.

Changed the comments in the `GlobalSettings` class from Chinese to English, and added `KillZombieTrigger`, `DeathTrigger`, and `AutoRestart` properties.

Added `EnableAutoZombieCleanup` and `AutoZombieCleanupThreshold` properties in the `GlobalSettings` class.
Added `OnModEventHubEntitySpawned` handler for `ModEventHub.EntitySpawned` event in `GlobalSettings` constructor.

Added `OnModEventHubEntitySpawned` method to check the number of zombies and perform cleanup operations when automatic zombie cleanup is enabled.

Modified the logic of the parameter `inMainThread` of the `Utils.ExecuteConsoleCommand` method so that it can be executed correctly even when it is not the main thread.

Modified the execution logic of the automatic restart message in the `GlobalSettings` class, and changed the `inMainThread` parameter of `Utils.ExecuteConsoleCommand` from `true` to `false`.