Updated settings and method patches, added chat color table
In the `functionsettings.json` file, changed the value of `BlockFamilySharingAccount` from `true` to `false` to allow family sharing accounts.

In the `GlobalSettings.cs` file, replaced the call of the `After_RequestToSpawnPlayer` method in the `GameManagerPatcher` class with the `Before_RequestToSpawnPlayer` method, and changed the patch type from `postfix` to `prefix`.

In the `GameManagerPatcher.cs` file, renamed the `After_RequestToSpawnPlayer` method to `Before_RequestToSpawnPlayer`.

In the `6-ColoredChat.sql` file, added a new table named `T_ColoredChat` to store the player's chat color information, including the `Id`, `CreatedAt`, `NameColor`, `TextColor`, and `Description` fields.