Added CustomName field and related processing logic
In the `6-ColoredChat.sql` file, a `CustomName` field was added to store custom names.

In the `T_ColoredChat.cs` file, a `CustomName` property was added and corresponding comments were added.

In the `ColoredChat.cs` file, the `SdtdServerKit.Models` namespace was imported, and a check was added in the `OnChatMessage` method to see if `managedPlayer` is `null`. After getting `coloredChat`, the processing logic for `CustomName` was added. If `CustomName` is not null, `CustomName` is used to render the player name.

In the `ColoredChatController.cs` file, the `System.Security.AccessControl` namespace was imported, and the processing of `CustomName` was added in the `Create` and `Update` methods.