Modified entity spawning logic and added remove entity command
In `GlobalSettings.cs`, the signature of the `OnEntitySpawned` method was modified to accept an `EntityInfo` parameter and use the `Utils.ExecuteConsoleCommand` command to delete the specific entity and log it when the automatic zombie cleanup is triggered.

In `ModEventHub.cs`, the signature of the `EntitySpawned` event was modified to accept an `EntityInfo` parameter and pass the `EntityInfo` parameter when calling the event in the `OnEntitySpawned` method.

In `Broadcaster.cs`, the handler of the `EntitySpawned` event was modified to accept and pass the `EntityInfo` parameter.

A new `RemoveEntity.cs` file was added, which defines a console command class called `RemoveEntity`, which implements the function of removing entities from the game. This class includes command description, help text, command list, and logic for executing commands.