Refactoring and extending the undo operation function
In the `RenderPrefab.cs` and `UndoPrefab.cs` files, the reference to `senderInfo.RemoteClientInfo` was removed, and the `senderInfo.GetEntityId()` method was used to obtain the `entityId`. In `UndoPrefab.cs`, the `GetMaxUndoHistorySize` method was added.

Added the `SetUndoSize.cs` file, which defines the `SetUndoSize` command, which is used to set or get the history size of the undo operation.

Added the `CommandSenderInfoExtension.cs` file, which defines the `GetEntityId` extension method, which is used to obtain the `entityId` from the `CommandSenderInfo` object.