Refactor Prefab related code, add paging and search functions
In the `AvailablePrefab.cs` file, rename the `PrefabName` property to `Name` and change its type to a nullable type `string?`. At the same time, add a new nullable type `string?` property `LocalizationName`. Change the type of the `FullPath` property to a nullable type `string?`.

In the `PrefabController.cs` file, remove unnecessary `using` statements. Change the return type of the `AvailablePrefabs` method from `IEnumerable<AvailablePrefab>` to `Paged<AvailablePrefab>`, and add paging and search functions.

In the `AvailablePrefabQuery.cs` file, add a new class `AvailablePrefabQuery` to handle paging and search query parameters.