Added CheckBlockType command class and related methods
Added a reference to `SdtdServerKit.Managers`.

Added a new `SdtdServerKit.Commands` namespace.

Added a new `CheckBlockType` class in this namespace, which inherits from `ConsoleCmdBase`.

The following methods are implemented in the `CheckBlockType` class:

* `getDescription`: Returns the description of the command.

* `getHelp`: Returns the help information of the command, including usage and parameter description.

* `getCommands`: Returns an array of aliases for the command.

* `Execute`: Executes the logic of the command, checks the type of the block at the specified position or under the player's feet according to the passed parameters, and outputs relevant information.

In the `Execute` method, handles the situation where the number of parameters is incorrect, and records the error information when catching the exception.