Updated multiple files to improve functionality and add new features
In the `CheckBlockType.cs` file, updated the comments to describe the functionality of the class.
In the `PlacePrefab.cs` file, changed the parameters of the `copyFromWorld` method from using a single coordinate value to using the sum of the offset position and the prefab size.
In the `UndoPrefab.cs` file, updated the parameters of the `SetUndo` method to add a `prefabInstanceId` parameter with a default value of -1.
In the `DuplicateArea.cs` file, added a new class called `DuplicateArea` that implements a console command for copying an area to another location. The class contains the command description, help information, command name, and execution logic. The execution logic includes parsing parameters, getting the player position, copying the area, handling rotation, updating blocks, etc.