Optimize blood moon calculation logic and refactor code
In the `DuplicateArea.cs` file, advance the acquisition of `entityId` and use the variable in multiple places, replacing the previous method of directly obtaining `entityId` from `senderInfo`.

In the `SkyChangeTrigger.cs` file, add the `bloodmoonFrequency` variable, and use `days % bloodmoonFrequency` to calculate `BloodMoonDaysRemaining` when `_isDark` is `true` and `false`, replacing the previous `Utilities.Utils.DaysRemaining(days)` method.

In the `Utils.cs` file, delete the `DaysRemaining` method, which is used to calculate the remaining days of the blood moon.