Removed old check for entities and added new remove method
Removed the check for `GameManager.Instance.World.Entities.dict` and the code to get the entity. Added a call to the `GameManager.Instance.World.RemoveEntity(entityId, EnumRemoveEntityReason.Despawned)` method to remove the entity. Removed the check for whether the entity is marked for unspawning and the corresponding logging. Added a new logging record to record that the entity has been removed.