Add SetPlayerCustomVar command class
Added a new namespace `SdtdServerKit.Commands` and added the `SetPlayerCustomVar` class. This class inherits from `ConsoleCmdBase` and implements the `getDescription()`, `getHelp()`, `getCommands()` and `Execute(List<string> args, CommandSenderInfo _senderInfo)` methods. The `Execute` method sets the player's custom variables based on the parameters and handles the cases of incorrect number of parameters and failure to obtain the player entity.