Refactor query logic and add ChatType support
In the `ChatRecordQueryDto.cs` file, added the `ChatRecordQueryDto` class, which inherits from the `PaginationQueryDto<ChatRecordQueryOrder>`, and added the `StartDateTime`, `EndDateTime`, and `ChatType` properties.

In the `DateTimeQueryDto.cs` file, deleted the `DateTimeQueryDto` class.

In the `IChatRecordRepository.cs` file, changed the parameter type of the `GetPagedListAsync` method from `DateTimeQueryDto` to `ChatRecordQueryDto`.

In the `ChatRecordRepository.cs` file, the parameter type of the `GetPagedListAsync` method was changed from `DateTimeQueryDto` to `ChatRecordQueryDto`, and the handling of `ChatType` was added to the query condition.

In the `ChatRecordQuery.cs` file, the class name was changed from `DateTimeQuery` to `ChatRecordQuery`, and the `ChatType` attribute was added.

In the `ChatType.cs` file, the `Unknown` enumeration value was added, and a comma was added after the `Whisper` enumeration value.

In the `ChatRecordController.cs` file, the parameter type of the `Get` method was changed from `DateTimeQuery` to `ChatRecordQuery`, and the `ChatType` attribute was added to the DTO mapping.