Add 100ms delay in processing loop
In the `GameStore.cs` and `VipGift.cs` files, add `await Task.Delay(100);` in the loops that process items and commands, respectively, to introduce a 100ms delay. This is probably to prevent performance issues or other side effects caused by performing a large number of operations in a short period of time.