Added error handling for language setting retrieval
Added a `try-catch` block in `ConfigManager.cs` to handle the process of getting language settings from `GamePrefs`. If the retrieval fails, the default English language will be used and a warning will be logged. The specific changes include adding a `string language;` variable declaration, using a `try-catch` block to try to get the language setting, if the retrieval fails, setting `language` to English and logging a warning, and finally passing `language` to the `GetRequired<GlobalSettings>` method.