Optimize delay time and improve code readability
In `GameStore.cs` and `VipGift.cs` files, reduce the delay time of `Task.Delay` from 100 milliseconds to 20 milliseconds.
In `PointsInfo.cs` file, initialize the `Id` property to `null!` and change the `PlayerName` property to a nullable type.
In `VipGiftController.cs` file, remove a redundant XML comment.
In `VipGiftController.cs` file, rename a variable `userTagRepository` to `repository` and simplify the code.