Added Task Scheduling Model and Controller, Updated Project References
In the `TaskScheduleModel.cs` file, added `TaskScheduleAdd`, `TaskScheduleUpdate`, and `TaskScheduleModel` classes to represent the add, update, and complete models of task scheduling.

In the `SdtdServerKit.csproj` file, added a reference to the `IceCoffee.Cron` package, and added the `NoWarn` property in the project properties to ignore specific warnings.

In the `TaskScheduleController.cs` file, added the `TaskScheduleController` controller class, which contains the functions to get, create, update, and delete task schedules, as well as the functions to get and update commands associated with task schedules.