Delete the IsRunning property and optimize exception handling
Deleted the `IsRunning` property in the `TaskScheduleModel.cs` file.
Deleted the setting of the `IsRunning` property in the `TaskScheduleController.cs` file.
Added `try-catch` blocks to the `Post` and `Put` methods to catch exceptions, and returned a `BadRequest` response when an exception is caught, and adjusted the indentation and structure of the code.
Deleted the code for getting the entity in the `Delete` method, directly called the `_taskSchedule.RemoveTaskSchedule(id)` method, and changed the return value of the delete operation to the number of deleted records.