Enhance Autofac integration with Web API
In the `Startup.cs` file, under the `SdtdServerKit.WebApi` namespace, a few lines of code were added to enhance the integration of Autofac and Web API.

Added `app.UseAutofacMiddleware(ModApi.ServiceContainer);` to register the Autofac middleware, supporting request scope and middleware injection.

Added `app.UseAutofacWebApi(config);` to integrate the Autofac container with Web API, set the dependency resolver of Web API to Autofac, and ensure that the Web API controllers and their dependencies are managed by the Autofac container.

Added `app.UseWebApi(config);` to add the runtime (middleware) of Web API to the OWIN pipeline, responsible for handling HTTP requests and routing them to the correct Web API controllers and action methods.