Update project path, modify method signatures and property definitions
Change the path of the `SdtdServerKit` project in the `7DaysToDie-ServerKit.sln` file from `"SdtdServerKit\SdtdServerKit.csproj"` to `"src\SdtdServerKit\SdtdServerKit.csproj"`.

In the `SdtdServerKit.Commands` namespace of the `SetPlayerColoredChat.cs` file, the signature of the `Execute` method was changed from `public override async void Execute` to `public override void Execute`, and the `async` keyword was removed.

In the `AutoBackupSettings.cs` file, the `ArchiveFolder` property was changed from `public string ArchiveFolder { get; set; } = null!;` to `public required string ArchiveFolder { get; set; }`, and the `required` keyword was added.