Add new namespace and update property type
In the `CdKeyRedeemVariables.cs` file, add a new namespace `SdtdServerKit.Variables`, and define an inner class `CdKeyRedeemVariables` in it. This class inherits from `VariablesBase` and adds a new nullable string property `CdKeyDescription` to describe the information of CdKey.

In the `VipGiftVariables.cs` file, change the `GiftName` property from a normal string type to a `required` string type. This means that when instantiating the `VipGiftVariables` class, you must assign a value to the `GiftName` property.