Added methods to get records and lists and updated comments
In the `ICdKeyRedeemRecordRepository.cs` file, added the `GetByKeyAndPlayerIdAsync` method for getting redemption records by key and player ID.

In the `ICdKeyRepository.cs` file, added the `GetByKeyAsync` method for getting Cd Key by key.

In the `ICommandListRepository.cs` file, updated the comments, changed the Chinese comments to English, and added the `GetListByCdKeyIdAsync` method for getting command lists by CdKey ID.

In the `IItemListRepository.cs` file, updated the comments, changed the Chinese comments to English, and added the `GetListByCdKeyIdAsync` method for getting item lists by CdKey ID.

In the `CdKeyRedeemRecordRepository.cs` file, the implementation of the `GetByKeyAndPlayerIdAsync` method is added to get the redemption record by key and player ID.

In the `CdKeyRepository.cs` file, the implementation of the `GetByKeyAsync` method is added to get the Cd Key by key.

In the `CommandListRepository.cs` file, the implementation of the `GetListByCdKeyIdAsync` method is added to get the command list by CdKey ID.

In the `ItemListRepository.cs` file, the implementation of the `GetListByCdKeyIdAsync` method is added to get the item list by CdKey ID.