Added POI area sleeping bag prompt function and cleaned up useless code
In the `functionsettings.json` file, deleted the `AutoRestart` configuration item, and added the `RemoveSleepingBagFromPoiTip` configuration item to remind players that they cannot place sleeping bags in the POI area.

In the `GlobalSettings.cs` file, added the `RemoveSleepingBagFromPoiTip` property to get or set the prompt information for removing sleeping bags in the POI area.

In the `GlobalSettings.cs` file, deleted some unnecessary `using` statements.

In the `GameManagerPatcher.cs` file, added the function of sending prompt information to players when detecting the placement of sleeping bags in the POI area.