Added API method for getting historical player IDs
In the `PlayersController.cs` file, added the `GettHistoryPlayerIds` method. This method uses the `HttpGet` attribute and specifies the route as `HistoryPlayerIds` through the `Route` attribute. The returned result is of type `IEnumerable<PlayerBase>`, which contains the player information obtained from `GameManager.Instance.persistentPlayers.Players`, including a combination string of the main ID, display name, entity ID, and native ID.